Paging: Refactor page size from usize to u64

+++ src/arch/aarch64/kernel/scheduler.rs
@@ -47,11 +47,12 @@ impl TaskStacks {
 		let user_stack_size = if size < KERNEL_STACK_SIZE {
 			KERNEL_STACK_SIZE
 		} else {
-			align_up!(size, BasePageSize::SIZE)
+			align_up!(size, BasePageSize::SIZE as usize)
 		};
 		let total_size = user_stack_size + DEFAULT_STACK_SIZE + KERNEL_STACK_SIZE;
-		let virt_addr = crate::arch::mm::virtualmem::allocate(total_size + 4 * BasePageSize::SIZE)
-			.expect("Failed to allocate Virtual Memory for TaskStacks");
+		let virt_addr =
+			crate::arch::mm::virtualmem::allocate(total_size + 4 * BasePageSize::SIZE as usize)
+				.expect("Failed to allocate Virtual Memory for TaskStacks");
 		let phys_addr = crate::arch::mm::physicalmem::allocate(total_size)
 			.expect("Failed to allocate Physical Memory for TaskStacks");
 
@@ -68,7 +69,7 @@ impl TaskStacks {
 		crate::arch::mm::paging::map::<BasePageSize>(
 			virt_addr + BasePageSize::SIZE,
 			phys_addr,
-			KERNEL_STACK_SIZE / BasePageSize::SIZE,
+			KERNEL_STACK_SIZE / BasePageSize::SIZE as usize,
 			flags,
 		);
 
@@ -76,7 +77,7 @@ impl TaskStacks {
 		crate::arch::mm::paging::map::<BasePageSize>(
 			virt_addr + KERNEL_STACK_SIZE + 2 * BasePageSize::SIZE,
 			phys_addr + KERNEL_STACK_SIZE,
-			DEFAULT_STACK_SIZE / BasePageSize::SIZE,
+			DEFAULT_STACK_SIZE / BasePageSize::SIZE as usize,
 			flags,
 		);
 
@@ -84,14 +85,16 @@ impl TaskStacks {
 		crate::arch::mm::paging::map::<BasePageSize>(
 			virt_addr + KERNEL_STACK_SIZE + DEFAULT_STACK_SIZE + 3 * BasePageSize::SIZE,
 			phys_addr + KERNEL_STACK_SIZE + DEFAULT_STACK_SIZE,
-			user_stack_size / BasePageSize::SIZE,
+			user_stack_size / BasePageSize::SIZE as usize,
 			flags,
 		);
 
 		// clear user stack
 		unsafe {
 			ptr::write_bytes(
-				(virt_addr + KERNEL_STACK_SIZE + DEFAULT_STACK_SIZE + 3 * BasePageSize::SIZE)
+				(virt_addr
+					+ KERNEL_STACK_SIZE + DEFAULT_STACK_SIZE
+					+ 3 * BasePageSize::SIZE as usize)
 					.as_mut_ptr::<u8>(),
 				0xAC,
 				user_stack_size,
@@ -130,7 +133,9 @@ impl TaskStacks {
 		match self {
 			TaskStacks::Boot(_) => VirtAddr::zero(),
 			TaskStacks::Common(stacks) => {
-				stacks.virt_addr + KERNEL_STACK_SIZE + DEFAULT_STACK_SIZE + 3 * BasePageSize::SIZE
+				stacks.virt_addr
+					+ KERNEL_STACK_SIZE + DEFAULT_STACK_SIZE
+					+ 3 * BasePageSize::SIZE as usize
 			}
 		}
 	}
@@ -139,7 +144,7 @@ impl TaskStacks {
 		match self {
 			TaskStacks::Boot(stacks) => stacks.stack,
 			TaskStacks::Common(stacks) => {
-				stacks.virt_addr + KERNEL_STACK_SIZE + 2 * BasePageSize::SIZE
+				stacks.virt_addr + KERNEL_STACK_SIZE + 2 * BasePageSize::SIZE as usize
 			}
 		}
 	}
@@ -154,7 +159,7 @@ impl TaskStacks {
 	pub fn get_interrupt_stack(&self) -> VirtAddr {
 		match self {
 			TaskStacks::Boot(stacks) => stacks.ist0,
-			TaskStacks::Common(stacks) => stacks.virt_addr + BasePageSize::SIZE,
+			TaskStacks::Common(stacks) => stacks.virt_addr + BasePageSize::SIZE as usize,
 		}
 	}
 
@@ -177,11 +182,11 @@ impl Drop for TaskStacks {
 
 				crate::arch::mm::paging::unmap::<BasePageSize>(
 					stacks.virt_addr,
-					stacks.total_size / BasePageSize::SIZE + 4,
+					stacks.total_size / BasePageSize::SIZE as usize + 4,
 				);
 				crate::arch::mm::virtualmem::deallocate(
 					stacks.virt_addr,
-					stacks.total_size + 4 * BasePageSize::SIZE,
+					stacks.total_size + 4 * BasePageSize::SIZE as usize,
 				);
 				crate::arch::mm::physicalmem::deallocate(stacks.phys_addr, stacks.total_size);
 			}
