+++ src/arch/x86_64/mm/virtualmem.rs
@@ -19,7 +19,7 @@ pub fn init() {
 pub fn allocate(size: usize) -> Result<VirtAddr, AllocError> {
 	assert!(size > 0);
 	assert_eq!(
-		size % BasePageSize::SIZE,
+		size % BasePageSize::SIZE as usize,
 		0,
 		"Size {:#X} is not a multiple of {:#X}",
 		size,
@@ -47,7 +47,7 @@ pub fn allocate_aligned(size: usize, alignment: usize) -> Result<VirtAddr, Alloc
 		alignment
 	);
 	assert_eq!(
-		alignment % BasePageSize::SIZE,
+		alignment % BasePageSize::SIZE as usize,
 		0,
 		"Alignment {:#X} is not a multiple of {:#X}",
 		alignment,
@@ -83,7 +83,7 @@ pub fn deallocate(virtual_address: VirtAddr, size: usize) {
 	);
 	assert!(size > 0);
 	assert_eq!(
-		size % BasePageSize::SIZE,
+		size % BasePageSize::SIZE as usize,
 		0,
 		"Size {:#X} is not a multiple of {:#X}",
 		size,
@@ -115,7 +115,7 @@ pub fn deallocate(virtual_address: VirtAddr, size: usize) {
 	);
 	assert!(size > 0);
 	assert_eq!(
-		size % BasePageSize::SIZE,
+		size % BasePageSize::SIZE as usize,
 		0,
 		"Size {:#X} is not a multiple of {:#X}",
 		size,
